The Refactoring, part 1

+++ core/core.c
@@ -433,8 +433,9 @@ long TryUpdatePixTree(ENGD *engd, TREE *estr) {
         printf(TXL_FAIL" %s\n", estr->path);
     else {
         turn = estr->epix->ainf.uuid ^ estr->turn;
-        printf("[%4ld%c%c%c] %s\n", estr->epix->ainf.uuid >> 2, (char)stat,
-              (turn & 2)? 'D' : 'U', (turn & 1)? 'L' : 'R', estr->path);
+        printf("[%4ld%c%c%c] %s\n", (long)(estr->epix->ainf.uuid >> 2),
+              (char)stat, (turn & 2)? 'D' : 'U', (turn & 1)? 'L' : 'R',
+               estr->path);
     }
     return ~0;
 }
